pyproject.toml
[project]
name = "python-captcha-solver"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "buptsad", email = "saddestboy@bupt.edu.cn" }
]
dependencies = [
    "opencv-python>=4.11.0.86",
    "drissionpage>=4.1.0.17",
    "python-dotenv>=1.0.1",
    "matplotlib>=3.10.1",
    "ipykernel>=6.29.5",
    "tqdm>=4.66.5",
    "numpy>=2.1.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]
readme = "README.md"
requires-python = ">= 3.13"
license = { text = "MITNFA" }

[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
